Refactored machines from person package to created machine package
